% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ChromExperiment.R
\name{annotations}
\alias{annotations}
\alias{annotations<-}
\alias{annotations,ChromExperiment-method}
\alias{annotations<-,ChromExperiment,GRanges-method}
\alias{annotations<-,ChromExperiment,NULL-method}
\title{Get or set genomic annotation}
\usage{
annotations(x, ...)

annotations(x, ...) <- value

\S4method{annotations}{ChromExperiment}(x, ...)

\S4method{annotations}{ChromExperiment,GRanges}(x, ...) <- value

\S4method{annotations}{ChromExperiment,NULL}(x, ...) <- value
}
\arguments{
\item{x}{A \code{\link{ChromExperiment}} object.}

\item{...}{Arguments passed to other methods.}

\item{value}{\itemize{
\item A \code{GRanges} object.
\item \code{NULL}. This will remove the existing \code{annotations}
}}
}
\description{
Methods to get or set a \code{\link[GenomicRanges:GRanges-class]{GRanges}} specifying the genomic
annotation in a \code{\link{ChromExperiment}}.
}
\seealso{
\code{\link[Signac]{Annotation}}
}
