% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromExperiment.R
\name{ChromExperiment-subset}
\alias{ChromExperiment-subset}
\alias{dimnames<-,ChromExperiment,list-method}
\alias{[,ChromExperiment,ANY,ANY,ANY-method}
\title{Subset ChromExperiment object}
\usage{
\S4method{dimnames}{ChromExperiment,list}(x) <- value

\S4method{[}{ChromExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\link{ChromExperiment}}.}

\item{value}{An object of a class specified in the S4 method signature.}

\item{i, j}{Indices specifying rows or columns to extract.}

\item{...}{Arguments passed to other methods.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the
lowest possible dimension.}
}
\description{
Deal with the dimensions and subset of \code{\link{ChromExperiment}} object.
These methods are generally identical to those inherited operations, with
additional manipulation for \code{fragments(x)}.
}
\examples{
csce <- load_example_csce()
csce

colnames(csce) <- paste0("Sorted_", colnames(csce))
csce
fragments(csce)

csce[1:10, 1:20]

csce[, sample(colnames(csce), 20)]

}
