% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{coerce-objects}
\alias{coerce-objects}
\alias{toSCE}
\alias{toSCME}
\alias{toSeurat}
\title{Read and write the OBJECT.json file}
\usage{
toSCE(object, ...)

toSCME(object, ...)

toSeurat(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other metheds.}
}
\description{
The \code{OBJECT.json} file provides the metadata of the object that is
represented by the corresponding directory.
}
