% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dietObj.R
\name{dietObj}
\alias{dietObj}
\alias{dietObj.SummarizedExperiment}
\alias{dietObj.SingleCellExperiment}
\alias{dietObj.ChromExperiment}
\alias{dietObj.MultiAssayExperiment}
\alias{dietObj.SingleCellMultiExperiment}
\title{Slim down a SingleCellExperiment-like object}
\usage{
dietObj(object, ...)

\method{dietObj}{SummarizedExperiment}(object, assays = NULL, verbose = TRUE, ...)

\method{dietObj}{SingleCellExperiment}(
  object,
  assays = NULL,
  reducedDims = NULL,
  colPairs = NULL,
  rowPairs = NULL,
  altExps = NULL,
  verbose = TRUE,
  ...
)

\method{dietObj}{ChromExperiment}(
  object,
  assays = NULL,
  reducedDims = NULL,
  colPairs = NULL,
  rowPairs = NULL,
  altExps = NULL,
  fragments = NULL,
  annotations = NULL,
  seqinfo = NULL,
  verbose = TRUE,
  ...
)

\method{dietObj}{MultiAssayExperiment}(object, experiments = NULL, verbose = TRUE, ...)

\method{dietObj}{SingleCellMultiExperiment}(
  object,
  experiments = NULL,
  reducedDims = NULL,
  colPairs = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}-like object}

\item{...}{Arguments passed to other metheds.}

\item{assays}{Select \code{\link[SummarizedExperiment]{assays}} specified
here. If is \code{NA}, remove all assays.}

\item{verbose}{Print progress.}

\item{reducedDims}{Keep a subset of dimension reductions specified here. If
is \code{NA}, remove all \code{\link[SingleCellExperiment]{reducedDims}}.}

\item{colPairs}{Keep a subset of \code{\link[SingleCellExperiment]{colPairs}}
specified here. If is \code{NA}, remove all \code{colPairs}.}

\item{rowPairs}{Similar as \code{colPairs}.}

\item{altExps}{Keep a subset of \code{\link[SingleCellExperiment]{altExps}}
specified here. If is \code{NA}, remove all \code{altExps}.}

\item{fragments}{Logical scalar specifying whether to keep
\code{\link{fragments}}.}

\item{annotations}{Logical scalar specifying whether to keep
\code{\link{annotations}}.}

\item{seqinfo}{Logical scalar specifying whether to keep
\code{\link{seqinfo}}.}

\item{experiments}{Select \code{\link[MultiAssayExperiments]{experiments}}
specified here. Cannot be \code{NA} (remove all assays).}
}
\description{
Keep only certain aspects of the \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object.
}
