% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MultiAssayExperiment.R
\name{experiment}
\alias{experiment}
\alias{experiment,MultiAssayExperiment,character-method}
\alias{experiment,MultiAssayExperiment,numeric-method}
\alias{experiment,MultiAssayExperiment,missing-method}
\title{Accessing single experiment data in MultiAssayExperiment}
\usage{
experiment(x, e, ...)

\S4method{experiment}{MultiAssayExperiment,character}(
  x,
  e,
  withPrimaries = FALSE,
  withColData = FALSE,
  mode = c("append", "replace"),
  ...
)

\S4method{experiment}{MultiAssayExperiment,numeric}(
  x,
  e,
  withPrimaries = FALSE,
  withColData = FALSE,
  mode = c("append", "replace"),
  ...
)

\S4method{experiment}{MultiAssayExperiment,missing}(
  x,
  e,
  withPrimaries = FALSE,
  withColData = FALSE,
  mode = c("append", "replace"),
  ...
)
}
\arguments{
\item{x}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} object}

\item{e}{Which experiment to fetch. Can be a integer index or an experiment
name. If missing, will fetch the first experiment by default.}

\item{...}{Arguments passed to other methods.}

\item{withPrimaries}{Whether or not to use primary names to replace the
original column names}

\item{withColData}{Whether or not to also fetch the associated
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}} of \code{x}.}

\item{mode}{String indicating how \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
metadata should be added or just replaced to the extracted experiment.
Passed to \code{\link[MultiAssayExperiment]{getWithColData}}.}
}
\value{
The extracted experiment object
}
\description{
Supplemental methods for accessing single experiment from
\code{\link[MultiAssayExperiment]{MultiAssayExperiment}} objects.
}
\seealso{
\code{\link[MultiAssayExperiment:MultiAssayExperiment-methods]{experiments}}
}
