% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/formatObj.R
\name{formatObj}
\alias{formatObj}
\alias{formatObj.SummarizedExperiment}
\alias{formatObj.SingleCellExperiment}
\alias{formatObj.ChromExperiment}
\title{Format an object}
\usage{
formatObj(object, ...)

\method{formatObj}{SummarizedExperiment}(object, ...)

\method{formatObj}{SingleCellExperiment}(object, ...)

\method{formatObj}{ChromExperiment}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other metheds.}
}
\description{
Format an object
}
