% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ChromExperiment.R
\name{fragments}
\alias{fragments}
\alias{fragments<-}
\alias{fragments,ChromExperiment-method}
\alias{fragments<-,ChromExperiment,IterableFragments-method}
\alias{fragments<-,ChromExperiment,Fragment-method}
\alias{fragments<-,ChromExperiment,character-method}
\alias{fragments<-,ChromExperiment,list_OR_List-method}
\alias{fragments<-,ChromExperiment,NULL-method}
\title{Get or set fragments data}
\usage{
fragments(x, ...)

fragments(x, ...) <- value

\S4method{fragments}{ChromExperiment}(x, ...)

\S4method{fragments}{ChromExperiment,IterableFragments}(x, ...) <- value

\S4method{fragments}{ChromExperiment,Fragment}(x, ...) <- value

\S4method{fragments}{ChromExperiment,character}(x, ...) <- value

\S4method{fragments}{ChromExperiment,list_OR_List}(x, ...) <- value

\S4method{fragments}{ChromExperiment,NULL}(x, ...) <- value
}
\arguments{
\item{x}{A \code{\link{ChromExperiment}} object.}

\item{...}{Passed to \code{\link{openFragments}}}

\item{value}{\itemize{
\item A list of \code{\link[BPCells:IterableFragments-methods]{IterableFragments}} or
\code{\link[Signac]{Fragment}} objects.
\item A single \code{\link[BPCells:IterableFragments-methods]{IterableFragments}}  or
\code{\link[Signac]{Fragment}} object.
\item A string specifying the path to a tabix-indexed fragments file.
\item \code{NULL}. This will remove all \code{fragments} data.
}}
}
\value{
\itemize{
\item \code{fragments}: Retrieve the stored \code{\link[BPCells:IterableFragments-methods]{IterableFragments}}.
\item \verb{fragments<-}: \code{x} with updated \code{fragments}
}
}
\description{
Methods to get or set fragments data for a \code{\link{ChromExperiment}}
object, to hold information needed for working with fragment files.
}
\note{
Currently we only support storing fragments as a single
\code{\link[BPCells:IterableFragments-methods]{IterableFragments}} object. If the input object inherits from
\code{\link[Signac]{Fragment}}, it will be converted into a
\code{\link[BPCells:IterableFragments-methods]{IterableFragments}} via \code{\link{openFragments}}.
}
\seealso{
\itemize{
\item \code{\link[BPCells:IterableFragments-methods]{IterableFragments}}
\item \code{\link[Signac]{Fragment}}
}
}
