% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getPath.R
\name{getPath}
\alias{getPath}
\alias{getPath.IterableMatrix}
\alias{getPath.IterableFragments}
\title{Fetch the file/directory path information for an on-disk object}
\usage{
getPath(object, ...)

\method{getPath}{IterableMatrix}(object, ...)

\method{getPath}{IterableFragments}(object, ...)
}
\arguments{
\item{object}{A disk-backed object.}

\item{...}{Arguments passed to other metheds.}
}
\value{
A \code{\link{data.frame}} with three columns:
\itemize{
\item \code{path}: Absolute path of the backend file/directory.
\item \code{type}: Type of the backend file/directory (dir, h5, ...).
\item \code{group} : Name of HDF5 group in the backend file.
}
}
\description{
Get the path and the link (for HDF5-backed) name for an on-disk object.
}
\details{
This function can be useful to check the file path of an object to avoid IO
conflicts.
}
\examples{
### For merged BPCells matrix
bpce <- load_example_sce()
getPath(counts(bpce))

### For merged BPCells fragments
cbpce <- load_example_csce()
getPath(counts(cbpce))
getPath(fragments(cbpce))

### For merged BPCells object
cbpce2 <- load_example_csce("unsorted")
cbpce <- merge(cbpce, cbpce2)
getPath(counts(cbpce))
getPath(fragments(cbpce))

}
