% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment.R
\name{getWithPrimaries}
\alias{getWithPrimaries}
\title{Extracting an assay with primaries as column names}
\usage{
getWithPrimaries(x, i)
}
\arguments{
\item{x}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} object.}

\item{i}{Which experiment to be fetched. Can be an integer or character.}
}
\value{
The extracted experiment object.
}
\description{
Change the column names of extracted experiment with primaries stored in the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}} of \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}.
}
