% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dumpObj.R, R/prepObj.R,
%   R/readObj.R, R/reloadObj.R, R/writeObj.R
\name{Obj-IO}
\alias{Obj-IO}
\alias{writeObj}
\alias{reloadObj}
\alias{prepObj}
\alias{dumpObj}
\alias{dumpObj.default}
\alias{dumpObj.list}
\alias{dumpObj.SimpleList}
\alias{dumpObj.data.frame}
\alias{dumpObj.DataFrame}
\alias{dumpObj.matrix}
\alias{dumpObj.DelayedArray}
\alias{dumpObj.IterableMatrix}
\alias{dumpObj.dgCMatrix}
\alias{dumpObj.IterableFragments}
\alias{dumpObj.Seqinfo}
\alias{dumpObj.GRanges}
\alias{dumpObj.LinearEmbeddingMatrix}
\alias{dumpObj.RangedSummarizedExperiment}
\alias{dumpObj.SingleCellExperiment}
\alias{dumpObj.ChromExperiment}
\alias{dumpObj.MultiAssayExperiment}
\alias{dumpObj.SingleCellMultiExperiment}
\alias{prepObj.default}
\alias{prepObj.LinearEmbeddingMatrix}
\alias{prepObj.Seqinfo}
\alias{prepObj.RangedSummarizedExperiment}
\alias{prepObj.SingleCellExperiment}
\alias{prepObj.ChromExperiment}
\alias{prepObj.MultiAssayExperiment}
\alias{prepObj.SingleCellMultiExperiment}
\alias{readObj}
\alias{reloadObj.default}
\alias{reloadObj.list}
\alias{reloadObj.SimpleList}
\alias{reloadObj.data.frame}
\alias{reloadObj.DataFrame}
\alias{reloadObj.matrix}
\alias{reloadObj.DelayedArray}
\alias{reloadObj.IterableMatrix}
\alias{reloadObj.IterableFragments}
\alias{reloadObj.dgCMatrix}
\alias{reloadObj.RangedSummarizedExperiment}
\alias{reloadObj.SingleCellExperiment}
\alias{reloadObj.ChromExperiment}
\alias{reloadObj.MultiAssayExperiment}
\alias{reloadObj.SingleCellMultiExperiment}
\alias{writeObj.default}
\alias{writeObj.list}
\alias{writeObj.SimpleList}
\alias{writeObj.matrix}
\alias{writeObj.dgCMatrix}
\alias{writeObj.IterableMatrix}
\title{IO for different data structures}
\usage{
writeObj(object, ...)

reloadObj(object, ...)

prepObj(object, ...)

dumpObj(object, ...)

\method{dumpObj}{default}(object, path, verbose = TRUE, ...)

\method{dumpObj}{list}(object, path, verbose = TRUE, class = "List", ...)

\method{dumpObj}{SimpleList}(object, path, verbose = TRUE, class = "List", ...)

\method{dumpObj}{data.frame}(
  object,
  path,
  verbose = TRUE,
  class = "DataFrame",
  file.name = "DataFrame.csv",
  idx.name = "_index",
  ...
)

\method{dumpObj}{DataFrame}(
  object,
  path,
  verbose = TRUE,
  class = "DataFrame",
  file.name = "DataFrame.csv",
  idx.name = "_index",
  ...
)

\method{dumpObj}{matrix}(
  object,
  path,
  verbose = TRUE,
  class = "HDF5Matrix",
  file.name = "matrix.h5",
  name = "matrix",
  with.dimnames = TRUE,
  ...
)

\method{dumpObj}{DelayedArray}(
  object,
  path,
  verbose = TRUE,
  class = "HDF5Matrix",
  file.name = "matrix.h5",
  name = "matrix",
  with.dimnames = TRUE,
  ...
)

\method{dumpObj}{IterableMatrix}(
  object,
  path,
  verbose = TRUE,
  dir.name = "matrix",
  class = "MatrixDir",
  ...
)

\method{dumpObj}{dgCMatrix}(
  object,
  path,
  verbose = TRUE,
  dir.name = "matrix",
  class = "dgCMatrix",
  ...
)

\method{dumpObj}{IterableFragments}(
  object,
  path,
  verbose = TRUE,
  dir.name = "fragments",
  class = "FragmentsDir",
  ...
)

\method{dumpObj}{Seqinfo}(object, path, verbose = TRUE, class = "Seqinfo", ...)

\method{dumpObj}{GRanges}(
  object,
  path,
  verbose = TRUE,
  file.name = "GRanges.gtf",
  class = "GRanges",
  ...
)

\method{dumpObj}{LinearEmbeddingMatrix}(object, path, verbose = TRUE, class = "LinearEmbeddingMatrix", ...)

\method{dumpObj}{RangedSummarizedExperiment}(
  object,
  path,
  verbose = TRUE,
  class = "RangedSummarizedExperiment",
  ...
)

\method{dumpObj}{SingleCellExperiment}(object, path, verbose = TRUE, class = "SingleCellExperiment", ...)

\method{dumpObj}{ChromExperiment}(object, path, verbose = TRUE, class = "ChromExperiment", ...)

\method{dumpObj}{MultiAssayExperiment}(object, path, verbose = TRUE, class = "MultiAssayExperiment", ...)

\method{dumpObj}{SingleCellMultiExperiment}(object, path, verbose = TRUE, class = "SingleCellMultiExperiment", ...)

\method{prepObj}{default}(object, ...)

\method{prepObj}{LinearEmbeddingMatrix}(object, ...)

\method{prepObj}{Seqinfo}(object, ...)

\method{prepObj}{RangedSummarizedExperiment}(object, ...)

\method{prepObj}{RangedSummarizedExperiment}(object, ...)

\method{prepObj}{ChromExperiment}(object, ...)

\method{prepObj}{MultiAssayExperiment}(object, ...)

\method{prepObj}{SingleCellMultiExperiment}(object, ...)

readObj(path, verbose = TRUE, ...)

\method{reloadObj}{default}(object, path, verbose = TRUE, ...)

\method{reloadObj}{list}(object, path, verbose = TRUE, ...)

\method{reloadObj}{SimpleList}(object, path, verbose = TRUE, ...)

\method{reloadObj}{data.frame}(object, path, verbose = TRUE, ...)

\method{reloadObj}{DataFrame}(object, path, verbose = TRUE, ...)

\method{reloadObj}{matrix}(object, path, verbose = TRUE, to.mem = FALSE, ...)

\method{reloadObj}{matrix}(object, path, verbose = TRUE, ...)

\method{reloadObj}{IterableMatrix}(object, path, verbose = TRUE, ...)

\method{reloadObj}{IterableFragments}(object, path, verbose = TRUE, ...)

\method{reloadObj}{dgCMatrix}(object, path, verbose = TRUE, to.mem = FALSE, ...)

\method{reloadObj}{RangedSummarizedExperiment}(object, path, verbose = TRUE, ...)

\method{reloadObj}{SingleCellExperiment}(object, path, verbose = TRUE, ...)

\method{reloadObj}{ChromExperiment}(object, path, verbose = TRUE, ...)

\method{reloadObj}{MultiAssayExperiment}(object, path, verbose = TRUE, ...)

\method{reloadObj}{SingleCellMultiExperiment}(object, path, verbose = TRUE, ...)

\method{writeObj}{default}(object, path, overwrite = FALSE, verbose = TRUE, ...)

\method{writeObj}{list}(object, path, overwrite = FALSE, verbose = TRUE, ...)

\method{writeObj}{SimpleList}(object, path, overwrite = FALSE, verbose = TRUE, ...)

\method{writeObj}{matrix}(object, path, overwrite = FALSE, verbose = TRUE, to.mem = FALSE, ...)

\method{writeObj}{dgCMatrix}(object, path, overwrite = FALSE, verbose = TRUE, to.mem = FALSE, ...)

\method{writeObj}{IterableMatrix}(object, path, overwrite = FALSE, verbose = TRUE, to.mem = FALSE, ...)
}
\arguments{
\item{object}{An SCE-like object.}

\item{...}{Arguments passed to other metheds.}

\item{path}{Path where \code{object} to be saved.}

\item{verbose}{Print progress.}

\item{to.mem}{Whether to keep the matrix in RAM. If \code{FALSE}, will use
the on-disk format:
\itemize{
\item \code{\link[BPCells:IterableMatrix-methods]{IterableMatrix}} for sparse matrix
\item \code{\link[HDF5Array]{HDF5Array}} for dense matrix
}}

\item{overwrite}{Whether or not to overwrite the existing \code{path}.}
}
\description{
Read or write assorted R objects into appropriate on-disk representations.
}
