% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10xH5.R
\name{read10xH5}
\alias{read10xH5}
\title{Read 10X hdf5 file}
\usage{
read10xH5(path, use.names = TRUE, use.BPCells = FALSE)
}
\arguments{
\item{path}{Path to HDF5 file]}

\item{use.names}{Label row names with feature names rather than ID numbers.}

\item{use.BPCells}{Whether or not to use \pkg{BPCells} to hold the matrix.}
}
\description{
Read count matrix from 10X CellRanger hdf5 file.
}
