% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10xMtx.R
\name{read10xMtx}
\alias{read10xMtx}
\title{Load in data from 10X}
\usage{
read10xMtx(path, gene.column = 2, cell.column = 1, use.BPCells = FALSE)
}
\arguments{
\item{path}{Directory containing the matrix.mtx, genes.tsv (or features.tsv),
and barcodes.tsv files provided by 10X.}

\item{gene.column}{Specify which column of genes.tsv or features.tsv to use
for gene names; default is 2}

\item{cell.column}{Specify which column of barcodes.tsv to use for cell
names; default is 1}

\item{use.BPCells}{Whether or not to use \pkg{BPCells} to hold the matrix.}
}
\description{
Read the sparse data matrix provided by 10X genomics.
}
