% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readH5AD.R
\name{readH5AD}
\alias{readH5AD}
\title{Read H5AD file}
\usage{
readH5AD(
  path,
  name = "/",
  layers = NULL,
  obsm = NULL,
  obsp = NULL,
  varp = NULL,
  uns = NULL,
  raw = NULL,
  use.BPCells = FALSE
)
}
\arguments{
\item{path}{A H5AD file}

\item{name}{Name of the HDF5 group where the AnnData is stored.}

\item{layers}{Which \code{layers} to read. If \code{NULL}, read all layers.}

\item{obsm}{Which \code{obsm} to read. If \code{NULL}, read all \code{obsm}.}

\item{obsp, varp}{Which \code{obsp} (\code{varp}) to read. If \code{NULL},
read all \code{obsp} (\code{varp}).}

\item{uns}{Which slots of \code{uns} data to read. If \code{NULL}, read all
slots of \code{uns}.}

\item{raw}{Whether or not to read the \code{raw} of AnnData. By default is
\code{TRUE}.}

\item{use.BPCells}{Whether or not to use \pkg{BPCells} to hold the matrix.}
}
\description{
Create a \code{\link[SingleCellExperiment]{SingleCellExperiment}} from a
H5AD file.
}
