% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment.R
\name{SCE-backend}
\alias{SCE-backend}
\alias{path,SingleCellExperiment-method}
\alias{show,SingleCellExperiment-method}
\title{Get the backend path for SingleCellExperiment}
\usage{
\S4method{path}{SingleCellExperiment}(object, ...)

\S4method{show}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}}

\item{...}{Additional arguments, for use in specific methods.}
}
\description{
Get the backend path for SingleCellExperiment
}
