% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rename.R
\name{setRownames}
\alias{setRownames}
\alias{setColnames}
\alias{set-dimnames}
\alias{setRownames,ANY,ANY-method}
\alias{setRownames,MultiAssayExperiment,list_OR_List-method}
\alias{setRownames,MultiAssayExperiment,character-method}
\alias{setColnames,ANY,ANY-method}
\alias{setColnames,MultiAssayExperiment,list_OR_List-method}
\alias{setColnames,MultiAssayExperiment,character-method}
\alias{setColnames,SingleCellMultiExperiment,character-method}
\title{Rename the dimensions}
\usage{
setRownames(x, new.names, ...)

setColnames(x, new.names, ...)

\S4method{setRownames}{ANY,ANY}(x, new.names, ...)

\S4method{setRownames}{MultiAssayExperiment,list_OR_List}(x, new.names, i, ...)

\S4method{setRownames}{MultiAssayExperiment,character}(x, new.names, i, ...)

\S4method{setColnames}{ANY,ANY}(x, new.names, ...)

\S4method{setColnames}{MultiAssayExperiment,list_OR_List}(x, new.names, i, ...)

\S4method{setColnames}{MultiAssayExperiment,character}(x, new.names, i, ...)

\S4method{setColnames}{SingleCellMultiExperiment,character}(x, new.names, i, ...)
}
\arguments{
\item{x}{A matrix-like object or object inheriting from
\code{\link[MultiAssayExperiment]{MultiAssayExperiment}}}

\item{new.names}{Strings to set as the new dimension names. Can be a single
vector of characters or a list containing multiple vectors.}

\item{...}{Arguments passed to other metheds.}

\item{i}{For a \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} object, should be
integers or characters specifying the experiment(s) to be renamed.
\itemize{
\item When \code{new.names} is a list, the length of \code{i} should be the same as
\code{new.names}. If \code{i} is \code{NA} or missing, all experiments will be used by
default.
\item When \code{new.names} is a character vector, \code{i} should also be a single
integer or character. If \code{i} is \code{NA} or missing, will change the primary
dimension names of \code{x}.
}}
}
\description{
Methods to set new dimension names to an object. These helpers are mostly
designed for \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} and its sub-classes.
}
