% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellMultiExperiment.R
\name{int_SCE}
\alias{int_SCE}
\alias{int_SCE<-}
\alias{SingleCellMultiExperiment-internal}
\alias{int_SCE,SingleCellMultiExperiment-method}
\alias{int_SCE<-,SingleCellMultiExperiment,SingleCellExperiment-method}
\alias{int_metadata,SingleCellMultiExperiment-method}
\alias{int_metadata<-,SingleCellMultiExperiment-method}
\title{Internal SingleCellExperiment object for SingleCellMultiExperiment class}
\usage{
int_SCE(x, ...)

int_SCE(x, ...) <- value

\S4method{int_SCE}{SingleCellMultiExperiment}(x)

\S4method{int_SCE}{SingleCellMultiExperiment,SingleCellExperiment}(x, ...) <- value

\S4method{int_metadata}{SingleCellMultiExperiment}(x)

\S4method{int_metadata}{SingleCellMultiExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{\link{SingleCellMultiExperiment}} object}

\item{...}{Arguments passed to other metheds.}

\item{value}{An object of a class specified in the S4 method signature.}
}
\description{
Methods to get or set the internal \code{\link[SingleCellExperiment]{SingleCellExperiment}}
slot in a \code{\link{SingleCellMultiExperiment}}. These functions are
intended for package development, and should not be used by ordinary users.
}
\details{
The internal \code{\link[SingleCellExperiment]{SingleCellExperiment}} is only used for storage
of \code{\link[SingleCellExperiment:SCE-internals]{int_metadata}} and \code{\link[SingleCellExperiment:SCE-internals]{int_colData}}.
}
