% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SingleCellMultiExperiment.R
\docType{class}
\name{SingleCellMultiExperiment-class}
\alias{SingleCellMultiExperiment-class}
\alias{SingleCellMultiExperiment}
\alias{show,SingleCellMultiExperiment-method}
\title{A single-cell MultiAssayExperiment class}
\usage{
SingleCellMultiExperiment(
  ...,
  reducedDims = list(),
  colPairs = list(),
  defaultExp = NULL
)

\S4method{show}{SingleCellMultiExperiment}(object)
}
\arguments{
\item{...}{Arguments passed to the constructor function of
\code{\link[MultiAssayExperiment]{MultiAssayExperiment}}, to fill the slots of the base class.}

\item{reducedDims, colPairs}{Arguments passed to the constructor of
\code{\link[SingleCellExperiment]{SingleCellExperiment}}. The manners of dimension alignments
are also the same.}

\item{defaultExp}{String defining the default \code{\link{experiment}}. If
\code{NULL}, will set the first one by default.}

\item{object}{A \code{SingleCellMultiExperiment}}
}
\value{
A \code{SingleCellMultiExperiment} object.
}
\description{
The \code{SingleCellMultiExperiment} class is designed to represent single-cell
multi-omics (or samples) data. The multiple-experiment part inherits from the
\code{\link[MultiAssayExperiment]{MultiAssayExperiment}} class and is used in the same manner.
In addition, this class supports storage of single-cell specific data fields,
such as \code{\link[SingleCellExperiment]{reducedDims}} and \code{\link[SingleCellExperiment]{colPairs}}.
}
\details{
In this class, the single-cell specific fields (\code{\link[SingleCellExperiment]{reducedDims}}
and \code{\link[SingleCellExperiment]{colPairs}}) are actually stored in \code{\link{int_SCE}}
slot, which is an internal \code{\link[SingleCellExperiment]{SingleCellExperiment}}. The column
names of \code{\link{int_SCE}} should be aligned with the row names of
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}} in the base \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
object.
}
