% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\name{sparse-NN}
\alias{sparse-NN}
\alias{sparse2NN}
\alias{nn2Sparse}
\alias{nn2Sparse.default}
\alias{nn2Sparse.Neighbor}
\alias{sparse2NN.dgCMatrix}
\alias{sparse2NN.dgTMatrix}
\title{Coerce between sparse distance matrix and nearest-neighbor object}
\usage{
sparse2NN(object, ...)

nn2Sparse(object, ...)

\method{nn2Sparse}{default}(object, repr = c("C", "T"), ...)

\method{nn2Sparse}{Neighbor}(object, repr = c("C", "T"), ...)

\method{sparse2NN}{dgCMatrix}(object, ...)

\method{sparse2NN}{dgTMatrix}(object, ...)
}
\arguments{
\item{object}{For \code{sparse2NN}, a \code{dgTMatrix} or \code{dgCMatrix}. For \code{nn2Sparse},
a \code{list} containing \code{idx} and \code{dist} or a \code{\link[NA]{Neighbor}}.}

\item{...}{Arguments passed to other metheds.}

\item{repr}{One of "C" and "T", specifying the representation of the sparse
matrix result}
}
\value{
\itemize{
\item \code{sparse2NN}: An NN \code{list} containing two matrix: \code{idx} and \code{dist}
\item \code{nn2Sparse}: A \code{dgCMatrix} or \code{dgTMatrix}
}
}
\description{
Coerce between sparse distance matrix and nearest-neighbor object
}
